+++ src/main/kotlin/klib/extensions/JsonObject.kt
@@ -1,0 +1,9 @@
+package klib.extensions
+
+import net.jemzart.jsonkraken.values.JsonObject
+
+inline fun <reified T> JsonObject.toObjectOfType(): T? {
+    if (this is T)
+        return this
+    return null
+}